+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-## Documentation Release
+## 2.3.0 (2025-09-07)
 
 ### ❤️ Released Documentation
 
-URL: https://eslint-airbnb-extended.nishargshah.dev
+- URL: https://eslint-airbnb-extended.nishargshah.dev
+
+### 🚨 Breaking Change
+
+- Requires `Node.js 16` or higher
+
+### 🩹 Fixes
+
+- **create-airbnb-x-config:** Fixed hook config issue
+- **create-airbnb-x-config:** The `Own Customization` option has been removed, since we now provide complete documentation to reference.
+- **eslint-config-airbnb-extended:** Fixed files issue ([#32](https://github.com/NishargShah/eslint-config-airbnb-extended/issues/32)).
+- **documentation:** Added changes of `CLI` in documentation.
 
 ## 2.2.0 (2025-08-25)